refactor: divided the xray_inject module into parts
