+++ io_scene_xray/ops/__init__.py
@@ -1,4 +1,8 @@
+
 import bpy
+
+from . import joint_limits
+
 
 class BaseOperator(bpy.types.Operator):
     report_catcher = None
